Don't log TransportErrors as errors
For our purposes, getting back a 4xx or 5xx isn't an actual error. It
just means that something is wrong with the request and they're going
to tell us all the details in the response. There's no need to log that
at error level. Instead we'll just log it as a warning. This has the
nice benefit of not getting sentry error messages anymore.